pyproject.toml
[project]
name = "speech-controller"
version = "0.1.0"
description = "speech-controller"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "boto3>=1.39.10",
    "numpy>=2.3.1",
    "openai>=1.97.0",
    "opencv-python>=4.11.0.86",
    "playsound==1.2.2",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "scipy>=1.16.0",
    "setuptools>=80.9.0",
    "sounddevice>=0.5.2",
    "webrtcvad>=2.0.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tool*", "componment*"]